feat: first step towards MVP
The first step contains a simple logic, which attempts to recover
original memory location of the values stored in registers. This logic
is triggered when an instruction is suspected to be a possible
transformation rule, e.g. ADD, IMUL, SUB etc.

The covered memory locations, if any, are thus treated as variables, and
will further be placed as nodes in the extracted DAG.